added object names and removed id arguments
removed "id" argument from constructors in favor of an optional string name system. this ensures non-dynamically allocated objects can consistently be found when refered to by files. (for now you can also do it dynamically with some difficulty).

also it could be because my battery is unplugged rn but it did seem a little laggier compared to the last successful build