improved logical consistency of ConnectorChain
getState and setState now work as you would expect if you were to change a sub-connector state directly. this means you dont have to reconstruct entire state vector in order to set state of sub-connector.  however work still needs to be done to rewrite getConnectorTransform because that only updates on setState